add page routes for lti13Config and lti13Jwks
